fix error message
